Add new icon for destinations page
Changes the icon used for selecting the frequent destinations on the map to a new one, like the ones used on the results page but grey instead of red or green.
Also renames the same icons of different colors that were incorrectly named "question_mark" despite not being one.
Fix: #30