Fix navigation
Fixes a bug related to the logic of the cars section. The section is skipped if the number of cars entered is 0, however this meant that the section is never marked as completed in that case, and the subsequents sections are greyed out and unskippable in the nav bar.
This commit changes the logic to mark the cars section as complete if there are 0 cars.
Also changes some generator made comments in the section configs that are incorrect.
Fix: #12